add viewport, add lang, remove type from style
Add viewport for proper formatting on mobile devices.
Add lang="en" to <html>, it helps browsers, screen readers etc to
interpret the website.
Remove text/css type from <style>, it is the default and specifying it
explicitly is considered deprecated in html5.
Specify fallback font so browser doesn't use a random font if Arial is
unavailable:
https://www.w3schools.com/Css/css_font_fallbacks.asp